+++ src/pages/Users/components/BookaSeat/index.jsx
@@ -1,0 +1,12 @@
+import Button from '@material-ui/core/Button';
+import WeekendOutlinedIcon from '@material-ui/icons/WeekendOutlined';
+
+const BookaSeat = () => {
+  return (
+    <Button title="Book a Seat" color="primary">
+      <WeekendOutlinedIcon />
+    </Button>
+  );
+};
+
+export default BookaSeat;
