+++ src/pages/Users/components/DeleteRow/index.jsx
@@ -1,0 +1,14 @@
+import Button from '@material-ui/core/Button';
+import DeleteOutlineIcon from '@material-ui/icons/DeleteOutline';
+
+const DeleteRow = () => {
+  const handleDeleteClick = () => {};
+
+  return (
+    <Button title="Delete" onClick={handleDeleteClick} color="secondary">
+      <DeleteOutlineIcon />
+    </Button>
+  );
+};
+
+export default DeleteRow;
