+++ src/pages/Users/components/GoToProfile/index.jsx
@@ -1,0 +1,12 @@
+import Button from '@material-ui/core/Button';
+import AccountCircleOutlinedIcon from '@material-ui/icons/AccountCircleOutlined';
+
+const GoToProfile = () => {
+  return (
+    <Button title="Profile" color="primary">
+      <AccountCircleOutlinedIcon />
+    </Button>
+  );
+};
+
+export default GoToProfile;
